feat: organise resets, swap in new element reset, fix heading size
