+++ lib/styles/reset/elements.css.ts
@@ -1,11 +1,10 @@
 import { style } from '@vanilla-extract/css';
 import { recipe, RecipeVariants } from '@vanilla-extract/recipes';
 
-import { overdriveTokens } from '../themes/theme.css';
-
-import { focusOutlineStyle } from './focusOutline.css';
-import { cssLayerReset } from './layers.css';
-import { sprinkles } from './sprinkles.css';
+import { overdriveTokens } from '../../themes/theme.css';
+import { focusOutlineStyle } from '../focusOutline.css';
+import { cssLayerReset } from '../layers.css';
+import { sprinkles } from '../sprinkles.css';
 
 export const trimmed = {
 	margin: 0,
